pick out a book from iCloud Drive
